fix: Made it also work on Korn Shell 93.
Before, when deciding the checksum mechanism, for example, it would try to compare "\$'1> /dev/null\n1> /dev/null\nopenssl'" to "'openssl'" and, obviously, fail; then, the solution was to redirect the standard output (&1) of "command -v $ccmd" to the standard error output (&2) and then redirect it to the null device (using 2>/dev/null).

Uma simples formalidade, as always.